// frontend/src/app/api/model/noticia.ts
/**
 * API oPROsita
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Noticia { 
    id?: number;
    descripcion?: string;
    archivoId?: number;
    grupoId?: number;
}